+++ eventbridge_event.go
@@ -1,4 +1,4 @@
-package main
+package cloudwatch_shipper
 
 import "time"
