+++ go.mod
@@ -4,14 +4,12 @@ go 1.18
 
 require (
 	github.com/aws/aws-lambda-go v1.34.1
-	github.com/stretchr/testify v1.8.0
+	github.com/aws/aws-sdk-go v1.44.91
 	go.uber.org/zap v1.23.0
 )
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	go.uber.org/atomic v1.9.0 // indirect
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
+	go.uber.org/atomic v1.7.0 // indirect
 	go.uber.org/multierr v1.6.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
