--- triggers/go.mod
@@ -1,15 +1,0 @@
-module main
-
-go 1.18
-
-require (
-	github.com/aws/aws-lambda-go v1.34.1
-	github.com/aws/aws-sdk-go v1.44.91
-	go.uber.org/zap v1.23.0
-)
-
-require (
-	github.com/jmespath/go-jmespath v0.4.0 // indirect
-	go.uber.org/atomic v1.7.0 // indirect
-	go.uber.org/multierr v1.6.0 // indirect
-)
