--- triggers/logger/logger.go
@@ -1,67 +1,0 @@
-package logger
-
-import (
-	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
-	"os"
-)
-
-const (
-	LogLevelDebug = "debug"
-	LogLevelInfo  = "info"
-	LogLevelWarn  = "warn"
-	LogLevelError = "error"
-	LogLevelFatal = "fatal"
-	LogLevelPanic = "panic"
-
-	envLogLevel = "LOG_LEVEL"
-
-	defaultLogLevel = LogLevelInfo
-)
-
-func GetLogger() *zap.Logger {
-	logLevel := getLogLevel()
-	cfg := zap.Config{
-		Encoding:         "json",
-		Level:            zap.NewAtomicLevelAt(logLevel),
-		OutputPaths:      []string{"stdout"},
-		ErrorOutputPaths: []string{"stderr"},
-		EncoderConfig: zapcore.EncoderConfig{
-			MessageKey:   "message",
-			LevelKey:     "level",
-			EncodeLevel:  zapcore.CapitalLevelEncoder,
-			TimeKey:      "time",
-			EncodeTime:   zapcore.ISO8601TimeEncoder,
-			CallerKey:    "caller",
-			EncodeCaller: zapcore.ShortCallerEncoder,
-		},
-	}
-	l, _ := cfg.Build()
-	return l
-}
-
-func getLogLevel() zapcore.Level {
-	logLevelStr := GetFuncLogLevel()
-	levelsMap := map[string]zapcore.Level{
-		LogLevelDebug: zapcore.DebugLevel,
-		LogLevelInfo:  zapcore.InfoLevel,
-		LogLevelWarn:  zapcore.WarnLevel,
-		LogLevelError: zapcore.ErrorLevel,
-		LogLevelPanic: zapcore.PanicLevel,
-		LogLevelFatal: zapcore.FatalLevel,
-	}
-
-	return levelsMap[logLevelStr]
-}
-
-func GetFuncLogLevel() string {
-	validLogLevels := []string{LogLevelDebug, LogLevelInfo, LogLevelWarn, LogLevelError, LogLevelFatal, LogLevelPanic}
-	logLevel := os.Getenv(envLogLevel)
-	for _, validLogLevel := range validLogLevels {
-		if validLogLevel == logLevel {
-			return validLogLevel
-		}
-	}
-
-	return defaultLogLevel
-}
